[#341] Fixing apigee_edge_apiproduct_rbac tests.
